# docs/swagger.yaml
definitions:
...
  gorm_model.JobRoleResp:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      sector:
        $ref: '#/definitions/gorm_model.SectorResp'
      sector_id:
        type: string
    type: object
  gorm_model.ProvinsiResp:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  gorm_model.SectorResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  gorm_model.UserResp:
    properties:
      account_status:
        type: string
      assigned_role_id:
        type: string
      bidang:
        type: string
      birthdate:
        type: string
      candidate_level:
        type: string
      email:
        type: string
      gpa:
        type: number
      id:
        type: string
      is_bookmark:
        type: boolean
      job_roles:
        items:
          $ref: '#/definitions/gorm_model.JobRoleResp'
        type: array
      job_title_id:
        type: string
      kabupaten_kota_id:
        type: string
      major:
        type: string
      must_reset_password:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      portofolio_link:
        type: string
      profile_picture:
        type: string
      province:
        $ref: '#/definitions/gorm_model.ProvinsiResp'
      recruitment_status_hex_code:
        type: string
      recruitment_status_id:
        type: string
      recruitment_status_name:
        type: string
      school_university:
        type: string
      summary:
        type: string
      system_role_name:
        type: string
      tech_stack:
        type: string
      unavailable_until:
        type: string
      years_experience:
        type: integer
    type: object
...
  request_model.ConfirmCVRequest:
    properties:
      edited_data:
        additionalProperties: true
        type: object
    required:
    - edited_data
    type: object
...
  request_model.CreateJobTitleRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      sector_id:
        type: string
    required:
    - name
    - sector_id
    type: object
...
  request_model.UpdateJobTitleRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      sector_id:
        type: string
    required:
    - name
    - sector_id
    type: object
...
  request_model.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  request_model.UpdateProfileRequest:
    properties:
      birthdate:
        description: Expecting YYYY-MM-DD
        type: string
      gpa:
        type: number
      job_role_ids:
        items:
          type: string
        type: array
      kabupaten_kota_id:
        type: string
      major:
        type: string
      name:
        type: string
      phone_number:
        type: string
      portofolio_link:
        type: string
      school_university:
        type: string
      summary:
        type: string
      tech_stack:
        items:
          type: string
        type: array
      years_experience:
        type: integer
    type: object
...
  response.Base:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      validation:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /cv/confirm:
    post:
      consumes:
      - application/json
      description: Confirm the parsed CV data after editing
      parameters:
      - description: Edited CV data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request_model.ConfirmCVRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Confirm parsed CV data
      tags:
      - CV
  /cv/generate:
    get:
      description: Get presigned CV download link for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Generate CV Link
      tags:
      - CV
  /cv/parsed:
    get:
      description: Fetch the parsed CV data by CV ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get parsed CV data
      tags:
      - CV
  /cv/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload CV
      parameters:
      - description: CV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Upload CV
      tags:
      - CV
...
  /job-titles:
    get:
      consumes:
      - application/json
      description: Get All Job Titles
      parameters:
      - description: Sector ID
        in: query
        name: sector_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get All Job Titles
      tags:
      - Job Title (Jabatan)
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.CreateJobTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Create Job Title
      tags:
      - Job Title (Jabatan)
  /job-titles/{id}:
    delete:
      parameters:
      - description: Title ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Delete Job Title
      tags:
      - Job Title (Jabatan)
    get:
      consumes:
      - application/json
      description: Get Job Title By ID
      parameters:
      - description: Title ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Job Title By ID
      tags:
      - Job Title (Jabatan)
    put:
      consumes:
      - application/json
      parameters:
      - description: Title ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdateJobTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Job Title
      tags:
      - Job Title (Jabatan)
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get detailed profile of current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gorm_model.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Update the profile of the current user
      parameters:
      - description: Profile update data
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - Profile
  /profile/password:
    put:
      consumes:
      - application/json
      description: Change the current user's password. Mandatory if must_reset_password
        is true.
      parameters:
      - description: Password update data
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Password
      tags:
      - Profile
  /profile/picture:
    post:
      consumes:
      - multipart/form-data
      description: Upload a profile picture for the current user
      parameters:
      - description: Profile picture file (JPEG, PNG, WebP)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Upload Profile Picture
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    description: 'Insert JWT Token with format: **Bearer {your token}**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"